Rename visualization modules
