+++ src/pylast/visualize/__init__.py
@@ -1,4 +1,4 @@
 from .visualize import *
-from .fast_visualizer import *
+from .event_visualizer import *
 from .lhaaso_background import *
-from .lact_root_quicklook import *
+from .event_quicklook import *
